+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## Version 1.78 - 2025-xx-xx
 
+* Better ReadMe
+
 
 ## Version 1.77 - 2025-03-31