+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # CHANGELOG
 
-## Version 1.78 - 2025-xx-xx
+## Version 1.78 - 2025-04-09
 
 * Allows to export table data
 * Fixes Labels wrongly shown in icons columns